Update index.html
CSS was applied on "dennis.jpeg. but I later made it a comment to temporary disable that feature. I did this so I'd be able to make new changes without creating another html file. The comments were done in lines 19 and 20 which had effect on the style (CSS) applied from lines 5 to 16.